RedfishClientPkg: introduce Redfish resource config2 protocol
Introduce Redfish resource config2 protocol. This is to add
additional parameter "JsonText" to Redfish resource config
protocol. Caller can provide JsonText to driver and driver does
not need to query BMC for data again. "Version" is added to
this protocol too so we don't create new protocol to add new
parameter in the future.

Signed-off-by: Nickle Wang <nicklew@nvidia.com>
Reviewed-by: Nick Ramirez <nramirez@nvidia.com>